// NextJS Quiz app Backend/src/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const Role = {
  User: 'User',
  Admin: 'Admin'
} as const

export type Role = (typeof Role)[keyof typeof Role]


export const Status = {
  Pending: 'Pending',
  Active: 'Active',
  Finished: 'Finished'
} as const

export type Status = (typeof Status)[keyof typeof Status]


export const Difficulty = {
  Easy: 'Easy',
  Medium: 'Medium',
  Hard: 'Hard'
} as const

export type Difficulty = (typeof Difficulty)[keyof typeof Difficulty]